chore: Remove tests package
