--- .env.template
@@ -1,2 +1,0 @@
-# development or production
-NODE_ENV=development
