+++ .eslintignore
@@ -9,16 +9,16 @@ node_modules
 .env
 .DS_Store
 .eslintrc.js
-vitest.config.mjs
+.lintstagedrc.js
+.prettierrc.js
+.stylelintrc.js
+vitest.units.config.mjs
+vitest.integration.config.ts
 /dist
 /cache
 /coverage
 /external
 /scripts
 /tests/mocks
-/tests/bin.mjs
 
-.lintstagedrc.js
-.prettierrc.js
-.stylelintrc.js
 esbuild/**/*.mjs
