+++ .github/workflows/release.yml
@@ -1,3 +1,8 @@
+# ⚠️ IMPORTANT: This workflow is a template from diplodoc/lint/scaffolding
+# Do not edit this file in other packages - it is automatically copied from here.
+# To modify this workflow, edit it in diplodoc/devops/lint/scaffolding/.github/workflows/
+# and then run the scaffolding update process.
+
 name: Release Package to npm
 
 # This workflow is configured in package.json publishConfig
@@ -27,9 +32,6 @@ jobs:
         with:
           node-version: 22
           registry-url: 'https://registry.npmjs.org'
-
-      - uses: codex-team/action-nodejs-package-info@v1
-        id: package
 
       - name: Install latest npm (>= 11.5.1)
         run: |
@@ -69,15 +71,5 @@ jobs:
       - name: Publish to npm
         if: github.event.release.draft == false || github.event_name == 'workflow_dispatch'
         run: npm publish --provenance --access public
-        env:
-          NODE_AUTH_TOKEN: ${{ secrets.NPM_TOKEN }}
-
-      - name: Publish tests to npm
-        working-directory: tests
-        if: github.event.release.draft == false || github.event_name == 'workflow_dispatch'
-        run: |
-          npm version ${{steps.package.outputs.version}}
-          npm ci
-          npm publish --provenance --access public
         env:
           NODE_AUTH_TOKEN: ${{ secrets.NPM_TOKEN }}
