--- tests/.npmignore
@@ -1,3 +1,0 @@
-output
-output-html
-output-static-html
