+++ tests/fixtures/cli.ts
@@ -1,7 +1,5 @@
-import type {Runner} from './runners';
-
-import {createRunner} from './runners';
-import {cleanupDirectory} from './utils/file';
+import runner from './runner';
+import {cleanupDirectory} from './file';
 
 export interface BuildRunArgs {
     md2md?: boolean;
@@ -17,14 +15,8 @@ export interface TranslateRunArgs {
 }
 
 class Build {
-    private readonly runner: Runner;
-
-    constructor(runner: Runner) {
-        this.runner = runner;
-    }
-
     run(input: string, output: string, args: string[]) {
-        return this.runner.runYfmDocs([
+        return runner.runYfmDocs([
             '--input',
             input,
             '--output',
@@ -37,14 +29,8 @@ class Build {
 }
 
 class Extract {
-    private readonly runner: Runner;
-
-    constructor(runner: Runner) {
-        this.runner = runner;
-    }
-
     run(input: string, output: string, args: string[]) {
-        return this.runner.runYfmDocs([
+        return runner.runYfmDocs([
             'translate',
             'extract',
             '--input',
@@ -58,11 +44,9 @@ class Extract {
 }
 
 export class CliTestAdapter {
-    readonly runner: Runner = createRunner();
+    readonly build = new Build();
 
-    readonly build = new Build(this.runner);
-
-    readonly extract = new Extract(this.runner);
+    readonly extract = new Extract();
 
     async testBuildPass(
         inputPath: string,
@@ -118,7 +102,7 @@ export class CliTestAdapter {
             ...additionalArgs.split(' ').filter(Boolean),
         ];
 
-        await this.runner.runYfmDocs(baseArgs);
+        await runner.runYfmDocs(baseArgs);
     }
 }
