+++ tests/fixtures/index.ts
@@ -1,3 +1,4 @@
-export * from './cliAdapter';
-export * from './utils/file';
-export * from './utils/test';
+export * from './cli';
+export * from './file';
+export * from './test';
+export * from './runner';
