+++ tests/fixtures/runner.ts
@@ -1,17 +1,18 @@
-import type {Runner} from './types';
-
+import {resolve} from 'node:path';
 import {execa} from 'execa';
 import strip from 'strip-ansi';
 
-export class BinaryRunner implements Runner {
-    private readonly binaryPath: string;
+export type Report = {
+    code: number;
+    warns: string[];
+    errors: string[];
+};
 
-    constructor(binaryPath: string) {
-        this.binaryPath = binaryPath;
-    }
+const bin = resolve(__dirname, '../../build/index.js');
 
+export class Runner {
     async runYfmDocs(argv: string[]) {
-        const {stderr, exitCode} = await execa(this.binaryPath, argv, {all: true, reject: false});
+        const {stderr, exitCode} = await execa(bin, argv, {all: true, reject: false});
         const report = {
             code: exitCode || 0,
             warns: fillLog(/^WARN/, stderr),
@@ -37,3 +38,5 @@ function fillLog(filter: RegExp, source: string) {
         .filter(Boolean)
         .filter((line) => line.match(filter));
 }
+
+export default new Runner();
