--- tests/fixtures/runners/types.ts
@@ -1,9 +1,0 @@
-export type Report = {
-    code: number;
-    warns: string[];
-    errors: string[];
-};
-
-export interface Runner {
-    runYfmDocs(argv: string[]): Promise<Report>;
-}
