--- tests/package.json
@@ -1,49 +1,0 @@
-{
-  "name": "@diplodoc/cli-tests",
-  "version": "0.0.0",
-  "bin": {
-    "diplodoc-cli-test": "bin.mjs"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git@github.com:diplodoc-platform/cli.git"
-  },
-  "exports": {
-    "./fixtures": "./fixtures/index.ts"
-  },
-  "files": [
-    "vitest.config.ts",
-    "tsconfig.json",
-    "bin.mjs",
-    "fixtures",
-    "mocks",
-    "e2e"
-  ],
-  "scripts": {
-    "build": "false",
-    "test": "vitest run",
-    "test:update": "vitest -u"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "devDependencies": {
-    "@diplodoc/cli": "file:.."
-  },
-  "dependencies": {
-    "@diplodoc/liquid": "^1.4.0",
-    "@types/node": "^22.19.0",
-    "@vitest/coverage-istanbul": "^3.1.1",
-    "@vitest/coverage-v8": "^3.1.1",
-    "execa": "^9.5.3",
-    "glob": "^11.0.2",
-    "js-yaml": "^4.1.0",
-    "strip-ansi": "^7.1.0",
-    "ts-dedent": "^2.2.0",
-    "typescript": "^5.9.2"
-  },
-  "peerDependencies": {
-    "@diplodoc/cli": "*",
-    "vitest": "^3.1.1"
-  }
-}
