+++ vitest.units.config.ts
@@ -10,9 +10,7 @@ export default defineConfig({
     test: {
         setupFiles: ['./src/test-setup.ts', './src/require.ts'],
         environment: 'node',
-        include: [
-            'src/**/*.{test,spec}.ts',
-        ],
+        include: ['src/**/*.{test,spec}.ts'],
         exclude: ['node_modules'],
         coverage: {
             all: true,
